+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All rights reserved.
 
 ### *Breaking changes*
 * Interface `TextTokenizer` removed. Use `TextAnalyzer.tokenize` and `TextAnalyzer.tokenizeJson` in stead.
+* Deleted mixin `LatinLanguageAnalyzerMixin`.
+* Moved class `TermSimilarityBase` from `text_analysis` library.
+* Moved all mixins and base-classes to `implementation` mini-library.
 
 ### *New*
 * New method `TextAnalyzer.tokenize`.